// src/app/models/loginUser/menuPermission.ts
export class MenuPermission {
    dashboard: boolean = false;
    contacts: boolean = false;
    pipeline: boolean = false;
    bulkImport: boolean = false;
    userManagement: boolean = false;
    settings: boolean = false;
}
export class ReadWritePermission {
    readPermission: boolean = false;
    writePermission: boolean = false;
}